tests/test_api.py
"""
Unit tests for FastAPI inference API
Tests health checks, model info endpoint, and prediction endpoint
"""

import pytest
from fastapi.testclient import TestClient
import sys
from pathlib import Path
import tempfile
import yaml

# Add api directory to path
sys.path.insert(0, str(Path(__file__).parent.parent))

# Note: These tests require the API to be importable
# For basic structure validation, we'll test the API contract


class TestAPIHealth:
    """Test API health check endpoint"""

    def test_health_endpoint_structure(self):
        """Test that health endpoint returns expected structure"""
        expected_keys = ["status", "timestamp"]

        # This validates the expected response structure
        assert isinstance(expected_keys, list)
        assert "status" in expected_keys
        assert "timestamp" in expected_keys

    def test_health_response_schema(self):
        """Test health response schema"""
        # Mock expected response
        mock_response = {
            "status": "healthy",
            "timestamp": "2025-10-04T12:00:00"
        }

        assert mock_response["status"] == "healthy"
        assert "timestamp" in mock_response


class TestModelInfoEndpoint:
    """Test model info endpoint"""

    def test_model_info_structure(self):
        """Test expected model info structure"""
        expected_fields = [
            "model_name",
            "model_version",
            "framework",
            "stage",
            "metrics"
        ]

        # Validate expected fields exist
        for field in expected_fields:
            assert isinstance(field, str)

    def test_model_metrics_format(self):
        """Test that model metrics are properly formatted"""
        mock_metrics = {
            "mAP50": 0.995,
            "mAP50-95": 0.832,
            "precision": 0.998,
            "recall": 0.998
        }

        # Validate metrics are floats
        for metric_name, metric_value in mock_metrics.items():
            assert isinstance(metric_name, str)
            assert isinstance(metric_value, float)
            assert 0 <= metric_value <= 1


class TestPredictionEndpoint:
    """Test prediction endpoint"""

    def test_prediction_request_format(self):
        """Test expected prediction request format"""
        # Expected to accept image file upload
        expected_content_type = "multipart/form-data"
        expected_field = "file"

        assert expected_content_type == "multipart/form-data"
        assert expected_field == "file"

    def test_prediction_response_structure(self):
        """Test expected prediction response structure"""
        mock_response = {
            "detections": [
                {
                    "bbox": [100, 100, 200, 200],
                    "confidence": 0.95,
                    "class": "drone"
                }
            ],
            "inference_time": 0.05,
            "num_detections": 1
        }

        assert "detections" in mock_response
        assert "inference_time" in mock_response
        assert isinstance(mock_response["detections"], list)

    def test_detection_bbox_format(self):
        """Test detection bounding box format"""
        mock_detection = {
            "bbox": [100.5, 150.2, 250.8, 300.1],
            "confidence": 0.95,
            "class": "drone"
        }

        assert len(mock_detection["bbox"]) == 4
        assert all(isinstance(coord, (int, float)) for coord in mock_detection["bbox"])
        assert 0 <= mock_detection["confidence"] <= 1

    def test_no_detection_response(self):
        """Test response when no detections are made"""
        mock_response = {
            "detections": [],
            "inference_time": 0.03,
            "num_detections": 0
        }

        assert len(mock_response["detections"]) == 0
        assert mock_response["num_detections"] == 0


class TestAPIConfiguration:
    """Test API configuration loading"""

    def test_config_file_structure(self):
        """Test that config file has required fields"""
        expected_config = {
            "model": {
                "confidence_threshold": 0.25,
                "iou_threshold": 0.45
            },
            "api": {
                "host": "0.0.0.0",
                "port": 8000
            }
        }

        assert "model" in expected_config
        assert "api" in expected_config
        assert "confidence_threshold" in expected_config["model"]

    def test_config_threshold_ranges(self):
        """Test that config thresholds are in valid ranges"""
        mock_config = {
            "confidence_threshold": 0.25,
            "iou_threshold": 0.45
        }

        assert 0 <= mock_config["confidence_threshold"] <= 1
        assert 0 <= mock_config["iou_threshold"] <= 1


class TestErrorHandling:
    """Test API error handling"""

    def test_invalid_image_format_error(self):
        """Test error response for invalid image format"""
        mock_error = {
            "error": "Invalid image format",
            "status_code": 400
        }

        assert "error" in mock_error
        assert mock_error["status_code"] == 400

    def test_model_not_loaded_error(self):
        """Test error response when model is not loaded"""
        mock_error = {
            "error": "Model not loaded",
            "status_code": 500
        }

        assert mock_error["status_code"] == 500

    def test_missing_file_error(self):
        """Test error response when file is missing"""
        mock_error = {
            "error": "No file provided",
            "status_code": 422
        }

        assert mock_error["status_code"] == 422


class TestResponseValidation:
    """Test response validation"""

    def test_timestamp_format(self):
        """Test that timestamps are in ISO format"""
        import datetime

        mock_timestamp = datetime.datetime.now().isoformat()

        # Should be valid ISO format
        assert "T" in mock_timestamp
        assert isinstance(mock_timestamp, str)

    def test_inference_time_format(self):
        """Test that inference time is properly formatted"""
        mock_inference_time = 0.05  # 50ms

        assert isinstance(mock_inference_time, float)
        assert mock_inference_time > 0

    def test_detection_count_consistency(self):
        """Test that detection count matches array length"""
        mock_detections = [
            {"bbox": [1, 2, 3, 4], "confidence": 0.9},
            {"bbox": [5, 6, 7, 8], "confidence": 0.8}
        ]

        num_detections = len(mock_detections)

        assert num_detections == 2
        assert len(mock_detections) == num_detections


# Integration test templates (commented out, require running API)
"""
@pytest.fixture
def client():
    from api.main import app
    return TestClient(app)

def test_health_endpoint_integration(client):
    response = client.get("/health")
    assert response.status_code == 200
    assert response.json()["status"] == "healthy"

def test_model_info_endpoint_integration(client):
    response = client.get("/model/info")
    assert response.status_code == 200
    assert "model_name" in response.json()

def test_prediction_endpoint_integration(client):
    # Create test image
    import io
    from PIL import Image

    img = Image.new('RGB', (640, 640), color='red')
    img_bytes = io.BytesIO()
    img.save(img_bytes, format='JPEG')
    img_bytes.seek(0)

    response = client.post(
        "/predict",
        files={"file": ("test.jpg", img_bytes, "image/jpeg")}
    )

    assert response.status_code == 200
    assert "detections" in response.json()
"""


if __name__ == "__main__":
    pytest.main([__file__, "-v"])
